feat: Python skeleton for code interview
## Context

The current process requires sending code samples in either Ruby and Go then asking the candidate to extend its functionality in the Pairing round. Near the end of 2021, the hiring panel at the time decided to explore the idea of providing a "skeleton" program for the candidates to fill in during the interview.

## What this is

This PR introduces a price calculator written as a Python CLI application with provided IO (JSON handling, stdout, etc.) but without the actual logic (which raises `NotImplementedError`). We will send them the requirements prior to their interview (the first two sections of README). The candidate's job is then to implement those pieces of logic and make sure the tests passed.